[pam_bt_pair] Split code from header
